+++ xarray/core/dataarray.py
@@ -6670,68 +6670,6 @@ class DataArray(
             **indexer_kwargs,
         )
 
-    def roundStringify(self, n: int) -> DataArray:
-        """
-        Converts rounded DataArray with specific significant figures to strings.
-
-        Parameters
-        ----------
-        n : int
-            The number of significant figures the DataArray would be converted to
-
-        Returns
-        -------
-        DataArray
-
-        Example
-        -------
-        Create a sample DataArray
-        >>> data = [1.234567, 2.345678, 3.456789]
-        >>> da = xr.DataArray(data, dims="x", name="my_data")
-        >>> rounded_da = da.roundStringify(3)
-
-        >>> da
-        <xarray.DataArray 'my_data' (x: 3)>
-        array([1.234567, 2.345678, 3.456789])
-        Dimensions without coordinates: x
-        >>> rounded_da
-        <xarray.DataArray 'None' (x: 3)>
-        array(['1.23', '2.35', '3.46'], dtype='<U4')
-        Dimensions without coordinates: x
-        """
-        # Handle edge cases
-        non_zero_vals = np.where(np.abs(self.values.astype(float)) == 0, False, True)
-
-        if not non_zero_vals.any():
-            new_name = f"{self.name}_rounded_to_{n}_sigfigs"
-            return DataArray(
-                np.full((self.size,), "0", dtype=str),
-                dims=self.dims,
-                name=new_name,
-                attrs=self.attrs,
-                coords=self.coords,
-            )
-        # Ensure non-negative precision
-        precision = max(
-            n
-            - 1
-            - np.floor(
-                np.log10(np.abs(self.values[non_zero_vals].astype(float))).astype(int)
-            ).any(),
-            0,
-        )
-        factor = 10**precision
-        rounded_arr = np.floor(self.values.astype(float) * factor) / factor
-        new_name = f"{self.name}_rounded_to_{n}_sigfigs"
-
-        return DataArray(
-            rounded_arr.astype(str),
-            dims=self.dims,
-            name=new_name,
-            attrs=self.attrs,
-            coords=self.coords,
-        )
-
     # this needs to be at the end, or mypy will confuse with `str`
     # https://mypy.readthedocs.io/en/latest/common_issues.html#dealing-with-conflicting-names
     str = utils.UncachedAccessor(StringAccessor["DataArray"])